backend/requirements.txt
asgiref==3.5.0
Django==4.0.4
djangorestframework==3.13.1
drf-nested-routers==0.93.4
pytz==2022.1
sqlparse==0.4.2
